Abstract Entity as an interface
